+++ src/components/common/fields/InputField.vue
@@ -14,23 +14,13 @@
     </div>
 
     <input
+      v-bind="$attrs"
+      v-on="listeners"
       class="input-field__input"
       :class="{ 'input-field__input--placeholder-auto-hidden': label }"
       :type="type"
-      :placeholder="placeholder || ' '"
+      :placeholder="$attrs.placeholder || ' '"
       :value="value"
-      :disabled="disabled"
-      :name="name"
-      :autocomplete="autocomplete"
-      :autofocus="autofocus"
-      :min="min"
-      :max="max"
-      :step="step"
-      :required="required"
-      :readonly="readonly"
-      :title="title"
-      :form="form"
-      @input="onInput"
       @focus="isInputFocused = true"
       @blur="isInputFocused = false"
     >
@@ -60,32 +50,19 @@ import InputFieldAutocomplete from './InputFieldAutocomplete'
 
 import 'mdi-vue/HelpCircleIcon'
 
+const EVENTS = {
+  input: 'input',
+}
+
 export default {
-  components: {
-    InputFieldAutocomplete,
-  },
+  components: { InputFieldAutocomplete },
 
   props: {
     label: { type: String, default: '' },
     value: { type: [String, Number], default: undefined },
     errorMessage: { type: String, default: undefined },
     autocompleteType: { type: String, default: '' },
-    // proxies
-    autocomplete: { type: String, default: 'off' },
-    autofocus: { type: Boolean, default: false },
-    disabled: { type: Boolean, default: false },
-    name: { type: String, default: undefined },
-    placeholder: { type: String, default: undefined },
     type: { type: String, default: 'text' },
-    required: { type: Boolean, default: true },
-    readonly: { type: Boolean, default: false },
-    title: { type: [String, Number], default: undefined },
-    form: { type: [String, Number], default: null },
-
-    // [type="number"] proxies
-    min: { type: [String, Number], default: undefined },
-    max: { type: [String, Number], default: undefined },
-    step: { type: [String, Number], default: undefined },
   },
 
   data () {
@@ -94,11 +71,20 @@ export default {
     }
   },
 
-  methods: {
-    onInput (event) {
-      this.beforeEmit(event.target)
-      this.$emit('input', event.target.value)
+  computed: {
+    listeners () {
+      return {
+        ...this.$listeners,
+        input: event => {
+          this.beforeEmit(event.target)
+          this.$emit(EVENTS.input, event.target.value)
+        },
+      }
     },
+  },
+
+  methods: {
+    onInput (event) {},
 
     beforeEmit (target) {
       if (this.type === 'number') {
@@ -119,8 +105,8 @@ export default {
 
     normalizeMinMax (target) {
       const { value } = target
-      const max = '' + this.max
-      const min = '' + this.min
+      const max = '' + this.$attrs.max
+      const min = '' + this.$attrs.min
 
       if (value === '') {
 
@@ -136,7 +122,7 @@ export default {
 
       let precision
       try {
-        precision = this.step.match(/(?:\.|,)\d+$/)[0].slice(1).length
+        precision = this.$attrs.step.match(/(?:\.|,)\d+$/)[0].slice(1).length
       } catch (error) {
         precision = 0
       }