+++ CHANGELOG.md
@@ -13,15 +13,25 @@ for further information about branching and tagging conventions.
 - Security asset type string to assetTypeToString filter
 - Postal code field in kyc request
 - Usage of Unix line endings rule
+- Vuelidate plugin for validating forms
 
 #### Changed
 - Align asset requests detail values to right side
 - Add datetime filter to kyc user's birthdate
 - Web client url in config changed to localhost:8060
+- Now using Vuelidate and form mixin instead of
+  native browser tools for validating forms
+- Extracted fees updating to a separate form
+- Extracted limits updating to a separate form
+- Now using $attrs property for input field proxies
+
+#### Removed
+- Vee-validate plugin
 
 #### Fixed
 - Show dash, if line 2 is not specified in kyc request
 - Terms field appearance in asset request
+- Error message styles on text field
 
 ## [1.7.0-rc.2] - 2019-05-21
 ### "Under the hood" changes